+++ block/events.go
@@ -1,0 +1,27 @@
+package block
+
+import (
+	"time"
+
+	coresequencer "github.com/rollkit/rollkit/core/sequencer"
+	"github.com/rollkit/rollkit/types"
+)
+
+// NewHeaderEvent is used to pass header and DA height to headerInCh
+type NewHeaderEvent struct {
+	Header   *types.SignedHeader
+	DAHeight uint64
+}
+
+// NewDataEvent is used to pass header and DA height to headerInCh
+type NewDataEvent struct {
+	Data     *types.Data
+	DAHeight uint64
+}
+
+// BatchData is used to pass batch, time and data (da.IDs) to BatchQueue
+type BatchData struct {
+	*coresequencer.Batch
+	time.Time
+	Data [][]byte
+}
